+++ packages/preset-panda/src/spacing.ts
@@ -1,4 +1,5 @@
 export const spacing = {
+  0: { value: '0rem' },
   0.5: { value: '0.125rem' },
   1: { value: '0.25rem' },
   1.5: { value: '0.375rem' },