+++ packages/studio/styled-system/types/system-types.d.ts
@@ -57,7 +57,7 @@ export type JsxStyleProps = SystemProperties &
     css?: SystemStyleObject
   }
 
-type Assign<T, U> = Omit<T, keyof U> & U
+export type Assign<T, U> = Omit<T, keyof U> & U
 
 export type PatchedHTMLProps = {
   htmlSize?: string | number