+++ sandbox/nuxt/nuxt.config.ts
@@ -1,5 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
+  telemetry: false,
   postcss: {
     plugins: { '@pandacss/dev/postcss': {} },
   },