+++ extension/vsix-builder/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @pandacss/vsix-builder
 
+## 0.4.0
+
 ## 0.3.2
 
 ## 0.3.1