+++ packages/studio/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pandacss/studio
 
+## 0.4.0
+
+### Patch Changes
+
+- d00eb17c: Add `auto` value where neccessary to base utilities.
+- Updated dependencies [c7b42325]
+- Updated dependencies [5b344b9c]
+  - @pandacss/types@0.4.0
+  - @pandacss/config@0.4.0
+  - @pandacss/node@0.4.0
+  - @pandacss/token-dictionary@0.4.0
+  - @pandacss/logger@0.4.0
+  - @pandacss/shared@0.4.0
+
 ## 0.3.2
 
 ### Patch Changes